applied swig.patch

+++ src/arkode/fmod_int32/farkode_lsrkstep_mod.c
@@ -500,14 +500,14 @@ SWIGEXPORT int _wrap_FLSRKStepGetMaxNumStages(void *farg1, int *farg2) {
 }
 
 
-SWIGEXPORT int _wrap_FLSRKStepGetNumRHSinDQ(void *farg1, int *farg2) {
+SWIGEXPORT int _wrap_FLSRKStepGetNumRHSinDQ(void *farg1, long *farg2) {
   int fresult ;
   void *arg1 = (void *) 0 ;
-  int *arg2 = (int *) 0 ;
+  long *arg2 = (long *) 0 ;
   int result;
   
   arg1 = (void *)(farg1);
-  arg2 = (int *)(farg2);
+  arg2 = (long *)(farg2);
   result = (int)LSRKStepGetNumRHSinDQ(arg1,arg2);
   fresult = (int)(result);
   return fresult;
